Finish

+++ Minitalk/Mandatory/sources/server.c
@@ -6,7 +6,7 @@
 /*   By: revieira <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/05 18:00:50 by revieira          #+#    #+#             */
-/*   Updated: 2023/01/07 12:42:18 by revieira         ###   ########.fr       */
+/*   Updated: 2023/01/07 16:10:39 by revieira         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -33,23 +33,47 @@ char	*add_char_to_str(char *str, char c)
 
 char	*print_msg(int pid, char *msg)
 {
-	ft_printf("%s\n", msg);
-	if (kill(pid, SIGUSR2) == -1)
+	if (!msg)
+	{
+		ft_printf("");
+		ft_printf("\n");
+		if (kill(pid, SIGUSR2) == -1)
+		{
+			if (msg)
+				free(msg);
+			ft_exit_program("Error: could not send signal to client", 1);
+		}
+	}
+	else
+	{
+		ft_printf("%s\n", msg);
+		if (kill(pid, SIGUSR2) == -1)
+		{
+			if (msg)
+				free(msg);
+			ft_exit_program("Error: could not send signal to client", 1);
+		}
+		free(msg);
+	}
+	return (NULL);
+}
+
+void	more_bits(int pid, char *msg)
+{
+	if (kill(pid, SIGUSR1) == -1)
 	{
 		if (msg)
 			free(msg);
 		ft_exit_program("Error: could not send signal to client", 1);
 	}
-	free(msg);
-	return (NULL);
 }
 
 void	signal_handler(int sig, siginfo_t *info, void *context)
 {
-	static int	bits_received;
-	static int	pid;
-	static char	c;
-	static char	*msg;
+	static int	bits_received = 0;
+	static int	pid = 0;
+	static char	c = 0;
+	static char	*msg = NULL;
 
 	context = NULL;
 	if (!pid)
@@ -57,8 +81,7 @@ void	signal_handler(int sig, siginfo_t *info, void *context)
 	if (sig == SIGUSR2)
 		c += (1 << bits_received);
 	bits_received++;
-	if (kill(pid, SIGUSR1) == -1)
-		ft_exit_program("Error: could not send signal to client\n", 1);
+	more_bits(pid, msg);
 	if (bits_received == 7)
 	{
 		if (c == 0)
@@ -73,24 +96,18 @@ void	signal_handler(int sig, siginfo_t *info, void *context)
 	}
 }
 
-void	inict_signals(struct sigaction *s_sa)
+int	main(void)
 {
-	sigset_t	sigset;
+	struct sigaction	s_sa;
+	sigset_t			sigset;
 
 	sigemptyset(&sigset);
 	sigaddset(&sigset, SIGUSR1);
 	sigaddset(&sigset, SIGUSR2);
-	s_sa->sa_handler = NULL;
-	s_sa->sa_mask = sigset;
-	s_sa->sa_flags = SA_SIGINFO;
-	s_sa->sa_sigaction = signal_handler;
-}
-
-int	main(void)
-{
-	struct sigaction	s_sa;
-
-	inict_signals(&s_sa);
+	s_sa.sa_handler = NULL;
+	s_sa.sa_mask = sigset;
+	s_sa.sa_flags = SA_SIGINFO;
+	s_sa.sa_sigaction = signal_handler;
 	sigaction(SIGUSR1, &s_sa, NULL);
 	sigaction(SIGUSR2, &s_sa, NULL);
 	ft_printf("Server PID: %d\n", getpid());
